requirements.txt
accelerate==0.17.0
bitsandbytes==0.40.1.post1
datasets==2.4.0
diffusers==0.17.0
huggingface-hub==0.14.1
keras==2.8.0
Keras-Preprocessing==1.1.2
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.7.0.72
pandas==2.0.2
pilgram==1.2.1
Pillow==9.5.0
protobuf==3.20.2
scipy==1.10.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
wandb==0.15.3
xformers==0.0.20
